+++ resolwe_bio/processes/import_data/sra_file.py
@@ -52,13 +52,13 @@ class ImportSra(Process):
     slug = "import-sra"
     name = "SRA data"
     process_type = "data:sra"
-    version = "1.1.0"
+    version = "1.2.0"
     category = "Import"
     scheduling_class = SchedulingClass.BATCH
     persistence = Persistence.TEMP
     requirements = {
         "expression-engine": "jinja",
-        "executor": {"docker": {"image": "resolwebio/common:1.5.0"}},
+        "executor": {"docker": {"image": "resolwebio/sra-tools:1.0.0"}},
         "resources": {
             "cores": 1,
             "memory": 1024,
@@ -139,7 +139,7 @@ class ImportSraSingle(Process):
     slug = "import-sra-single"
     name = "SRA data (single-end)"
     process_type = "data:reads:fastq:single"
-    version = "1.1.1"
+    version = "1.2.0"
     category = "Import"
     scheduling_class = SchedulingClass.BATCH
     persistence = Persistence.RAW
@@ -149,7 +149,7 @@ class ImportSraSingle(Process):
     }
     requirements = {
         "expression-engine": "jinja",
-        "executor": {"docker": {"image": "resolwebio/common:1.5.0"}},
+        "executor": {"docker": {"image": "resolwebio/sra-tools:1.0.0"}},
         "resources": {
             "cores": 1,
             "memory": 1024,
@@ -290,7 +290,7 @@ class ImportSraPaired(Process):
     slug = "import-sra-paired"
     name = "SRA data (paired-end)"
     process_type = "data:reads:fastq:paired"
-    version = "1.1.0"
+    version = "1.2.0"
     category = "Import"
     scheduling_class = SchedulingClass.BATCH
     persistence = Persistence.RAW
@@ -300,7 +300,7 @@ class ImportSraPaired(Process):
     }
     requirements = {
         "expression-engine": "jinja",
-        "executor": {"docker": {"image": "resolwebio/common:1.5.0"}},
+        "executor": {"docker": {"image": "resolwebio/sra-tools:1.0.0"}},
         "resources": {
             "cores": 1,
             "memory": 1024,